Normalized benefit

+++ src/main/java/se/kth/jabeja/annealing/AnnealingStrategy.java
@@ -2,6 +2,6 @@ package se.kth.jabeja.annealing;
 
 public interface AnnealingStrategy {
     float getTemperature();
-    boolean accept(int oldBenefit, int newBenefit);
+    boolean accept(double oldBenefit, double newBenefit);
     void anneal();
 }
