+++ src/main/java/se/kth/jabeja/annealing/LinearAnnealing.java
@@ -16,7 +16,7 @@ public class LinearAnnealing implements AnnealingStrategy {
     }
 
     @Override
-    public boolean accept(int oldBenefit, int newBenefit) {
+    public boolean accept(double oldBenefit, double newBenefit) {
         return newBenefit * getTemperature() > oldBenefit;
     }
