+++ src/main/java/se/kth/jabeja/annealing/RestartingLinearAnnealing.java
@@ -18,7 +18,7 @@ public class RestartingLinearAnnealing implements AnnealingStrategy {
     }
 
     @Override
-    public boolean accept(int oldBenefit, int newBenefit) {
+    public boolean accept(double oldBenefit, double newBenefit) {
         return newBenefit * getTemperature() > oldBenefit;
     }
